#271 combined cdi,guice and complex example into one
