+++ examples/mini-examples/welcome-example/src/main/java/de/saxsys/jfx/exampleapplication/model/Gender.java
