+++ examples/mini-examples/welcome-example/src/main/resources/de/saxsys/jfx/exampleapplication/view/maincontainer/MainContainerView.fxml
