+++ examples/mini-examples/welcome-example/src/main/resources/de/saxsys/jfx/exampleapplication/view/personlogin/PersonLoginView.fxml
