+++ examples/mini-examples/welcome-example/src/main/resources/de/saxsys/jfx/exampleapplication/view/personwelcome/PersonWelcomeView.fxml
