+++ examples/mini-examples/welcome-example/src/test/java/de/saxsys/jfx/viewmodel/personlogin/PersonLoginViewModelTest.java
