--- examples/mvvmfx-cdi-starter/README.md
@@ -1,10 +1,0 @@
-mvvmfx-cdi-starter
-========
-
-This contains the startup code to run the example application with the dependency injection framework CDI/Weld and the mvvmfx extension
-[mvvmfx-cdi](/mvvmfx-parent/mvvmfx-cdi).
-
-The main code of the example is implemented in [mvvmfx-complex-example](/examples/mvvmfx-complex-example).
-
-An alternative way to run the example application is to use Guice as dependency injection framework which can be seen in the
-[mvvmfx-guice-starter example](/examples/mvvmfx-guice-starter).
