--- examples/mvvmfx-guice-starter/README.md
@@ -1,10 +1,0 @@
-mvvmfx-guice-starter
-========
-
-This contains the startup code to run the example application with the dependency injection framework Guice and the mvvmfx extension
-[mvvmfx-guice](/mvvmfx-parent/mvvmfx-guice).
-
-The main code of the example is implemented in [mvvmfx-complex-example](/examples/mvvmfx-complex-example).
-
-An alternative way to run the example application is to use CDI/Weld as dependency injection framework which can be seen in the
-[mvvmfx-cdi-starter example](/examples/mvvmfx-cdi-starter).
